# src/main/resources/api.yaml
swagger: "2.0"
info:
  description: "jobad reader api"
  version: "1.0.0"
  title: "Jobad Reader Api"
  termsOfService: "tbc"
  contact:
    email: "apiteam@swagger.io"
basePath: "/api/v1/"
schemes:
- "https"
- "http"

paths:
  /jobad/{id}:
    get:
      tags:
      - "JobAd"
      summary: "find jobad by id"
      description: ""
      operationId: "findJobAdById"
      produces:
      - "application/json"
      parameters:
      - in: "path"
        name: "id"
        description: "id of jobad"
        required: true
        type: "string"
      responses:
        200:
          description: get states ok
          schema:
            $ref: "#/definitions/JobAdDetailResponse"
  
  /jobad/list:
    get:
      tags:
      - "JobAd"
      summary: "get jobad in list order by amend time with limit and page size"
      description: ""
      operationId: "findJobAds"
      produces: 
      - "application/json"
      parameters: 
      - in: "query"
        name: "size"
        description: "limit size of the items, min is 1, max is 100, default is 25"
        required: false
        default: 25
        type: integer
        minimum: 1
        maximum: 100
      - in: "query"
        name: "page"
        description: "page no. of the query"
        required: false
        default: 1
        type: integer
        minimum: 1
      responses:
        200:
          description: OK
          schema:
            $ref: "#/definitions/JobAdListResponse"
  
  /batch/update-jobads:
    get:
      tags: 
      - "Batch Job"
      summary: "fetch rss from rss reader, and update jobad list in db"
      description: ""
      operationId: "updateJobAds"
      produces: 
      - "application/json"
      responses:
        200:
          description: OK
          schema:
            $ref: "#/definitions/BaseResponse"

definitions:
  BaseResponse:
    type: "object"
    properties:
      code:
        type: "string"
      message:
        type: "string"

  JobAdListResponse:
    allOf: 
    - type: "object"
    - $ref: "#/definitions/BaseResponse"
    - properties:
        totalPageNum:
          type: integer
        jobAdItems:
          type: array
          items:
            $ref: "#/definitions/JobAdItem"

  JobAdDetailResponse:
    allOf: 
    - type: "object"
    - $ref: "#/definitions/BaseResponse"
    - properties:
        jobAdItem:
          $ref: "#/definitions/JobAdItem"

##########
# model
##########
  JobAdItem:
    properties:
      id:
        type: "string"
      companyName:
        type: "string"
      title:
        type: "string"
      industry:
        type: "string"
      careerLevel:
        type: "string"
      qualification:
        type: "string"
      location:
        type: "string"
      employmentType:
        type: "string"
      others:
        type: "string"
      yearsOfExp:
        type: "integer"
      benefits:
        type: "array"
        items:
           type: "string"
      postedDate:
        type: "string"
        format: "date"
      fromWebSite:
        type: "string"
      url:
        type: "string"

externalDocs:
  description: "Find out more about Swagger"
  url: "http://swagger.io"